+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Backwards incompatible changes
     - Minimum required elixir version is now "~> 1.5"
+    - Support for cayley version `0.6.1` has been dropped
 
 ## v0.8.0 (2018-09-09)